refactor : 커스텀 뷰 생성 및 뷰 레이아웃 리팩터링
