+++ Expo1900/Expo1900/View/ExpositionPosterChildViews/InformationLabel.swift
@@ -1,0 +1,33 @@
+//
+//  InformationLabel.swift
+//  Expo1900
+//
+//  Copyright (c) 2022 Minii All rights reserved.
+
+import UIKit
+
+class InformationLabel: UILabel {
+    init() {
+        super.init(frame: .zero)
+        textAlignment = .center
+    }
+    
+    init(aligment: NSTextAlignment, settingFont: UIFont?, lines: Int = 1) {
+        super.init(frame: .zero)
+        
+        textAlignment = aligment
+        
+        
+        if let settingFont {
+            font = settingFont
+        }
+        numberOfLines = lines
+    }
+    
+    
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        
+        textAlignment = .center
+    }
+}
